Updated version of the plugin now released.
I have added the /UpdatePlaytime <name> or /UpdatePlaytime (if its yourself) command which tries to update the specified player or your own playtime if there is a great enough change. I have also added the /UpdateAllPlaytimes command which updates all the online players play times. 
There is also 2 new permissions to go with this update, the PlaytimeFixer.UpdatePlaytime permission which allows you to use the /UpdatePlaytime command and the PlaytimeFixer.UpdateAllPlaytimes permission which allows you to use the /UpdateAllPlaytimes command.