Offline players and new cosmetics
Added the ability to update offline players playtime and also changed
around some of the messages sent to the user and target to include
successful/unsuccessful responses from the plugin. Added the time
changed into the success message recieved by the user when using the
commands as well and changed around the /updateallplaytimes confirmation
message.